utils.py
import torch
import torch.nn.functional as F

def compute_loss(outputs, labels):
    return F.cross_entropy(outputs, labels)

def compute_metrics(preds, labels):
    from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
    accuracy = accuracy_score(labels, preds)
    precision = precision_score(labels, preds, average='weighted')
    recall = recall_score(labels, preds, average='weighted')
    f1 = f1_score(labels, preds, average='weighted')
    return {"accuracy": accuracy, "precision": precision, "recall": recall, "f1": f1}
